queuer/requirements-pypi.txt
flask>=2.0.2,<3.0.0
google-cloud-pubsub>=2.11.0,<3.0.0
gunicorn>=20.1.0,<21.0.0